Decay

Documentation for Decay.

Decay.lambda

Decay.Becquerel

Decay.Curie

Decay.DPM

Decay.PerYear

Decay.RadUnit

Decay.D

Decay.D_

Decay.N

Decay.N2

Decay.Q

Decay.R

Decay.act2conc

Decay.conc2act

Decay.dN2_dt

Decay.decayconstant

Decay.halflife

Decay.lambda — Constant

Decay.λ
Decay.lambda

A NamedTuple library containing the decay constants and uncertainties of a selection of useful radiochronometric isotopes.

Decay constant central tendencies (typically means) are returned by providing the element (e.g. U) and isotope (e.g. 238), as in λ.U238. To call the 1s uncertainty (standard error of the mean), give the same value with a s prefix, as in λ.sU238.

In most cases, I try to cite the original publications containing accepted measurements. In a few cases, I use compilations from reviews.

isotope(s)	reference	doi	note

²³⁸U, ²³⁵U	Jaffey+ 1991	10.1103/PhysRevC.4.1889

²³²Th	Holden 1990	10.1351/pac199062050941	[1]

²³⁴U, ²³⁰Th	Cheng+ 2013	10.1016/j.epsl.2013.04.006

²²⁶Ra	Holden 1990	10.1351/pac199062050941

⁴⁰K	Renne+ 2011	10.1016/j.gca.2011.06.021	[2]

²³³U	Wikipedia	Isotopes of Uranium

²³⁶U	Wikipedia	Isotopes of Uranium

²²⁹Th	compilation	SRM 4328d certificate

[1] This value is similar to the often-used value of Le Roux & Glendenin 1963 (1.40 ± 0.007 Gy). Audi+ 1997 ( doi: 10.1016/S0375-9474(97)00482-X) reports 14.05 ± 0.04 Gy and references Bonetti+ 1995 (doi: [https://doi.org/10.1103/PhysRevC.51.2530, which does not appear to report a λ for α-decay.

[2] K40 and sK40 are NamedTuples containing fields β and e for the branched decay. The total decay constant is obtained with sum(λ.K40). I also include the now deprecated decay constants of Steiger & Jager 1977 (doi: 10.1016/0012-821X(77)90060-7). See also Renne+ 2010 (doi: 10.1016/j.gca.2010.06.017).

Decay.Becquerel — Type

Becquerel

A struct denoting becquerels, the SI unit of radioactivity, defined as an activity of one disintegration per second (s⁻¹). Learn more: https://en.wikipedia.org/wiki/Becquerel. Also converts any RadUnit to a Bq.

julia> Becquerel(Curie(1))
Becquerel(3.7e10)

Decay.Curie — Type

Curie

A struct denoting curies, a non-SI unit of radioactivity named in honor of Marie Skłodowska-Curie. Originally defined by radon emnation from a gram of radium, the curie (Ci) is now defined 1 Ci = 3.7 × 10¹⁰ Bq. Learn more: https://en.wikipedia.org/wiki/Curie_(unit).

Also converts any RadUnit to a Ci.

julia> Curie(Becquerel(3.7e10))
Curie(1.0)

Decay.DPM — Type

DPM

A struct denoting disintegrations per minute (dpm), a non-SI unit of radioactivity.

Also converts any RadUnit to a dpm.

julia> DPM(Becquerel(1))
DPM(60.0)

Decay.PerYear — Type

PerYear

A struct denoting disintegrations per year, a non-SI unit of radioactivity.

Also converts any RadUnit to a y⁻¹.

julia> PerYear(Curie(1))
PerYear(1172.459249119071)

Decay.RadUnit — Type

RadUnit

Supertype for all radioactivity units. Any RadUnit is callable by its single field n or by calling it as a function (see examples below). Converting between types is accomplished by inputting any RadUnit type into the constructor of another RadUnit type.

See also: Becquerel, Curie, DPM, PerYear.

Examples

julia> r = Becquerel(3)
Becquerel(3.0)

julia> r.n, r()
(3.0, 3.0)

Decay.D — Method

Decay.D(t, λ; N=1, Do=0)

The classic age equation.

Calculates the abundance of stable radiogenic daughter isotope D after time t, given the MODERN abundance of parent isotope N (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + N ( exp(-λ t) - 1 )

see also: D_

Decay.D_ — Method

Decay.D_(t, λ; No=1, Do=0)

Forward model a single decay system.

Calculates the abundance of radiogenic daughter isotope D after time t, given the INITIAL abundance of parent isotope No (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + Nₒ ( 1 - exp(-λ t) )

see also: D

Decay.N — Method

Decay.N(t, λ; No=1)

The classic decay equation. Calculate the remaining abundance of parent radioisotope N with decay constant λ after time t. Optionally provide an initial abundance No (default = 1).

Equation: N = Nₒ exp(-λ t)

Decay.N2 — Method

Decay.N2(t, λ₂, λ₁; N2o=0, N1o=1)

Calculate the the abundance of a radioactive daughter isotope in a decay-series after some time t, given its decay constant λ₂ and that of its parent λ₁. Optionally provide the initial abundances of the isotope N2o (default = 0) and its parent N1o (default = 1).

Equation: N₂ = (λ₁ / (λ₂ - λ₁)) * N₁⁰ * (exp(-λ₁ * t) - exp(-λ₂ * t)) + N₂ᵒ * exp(-λ₂ * t)

MathTex: N_2 = \frac{\lambda_1}{\lambda_2 - \lambda_1} N^0_1 \left( e^{-\lambda_1 t} - e^{-\lambda_2 t} \right) + N^0_2 e^{-\lambda_2 t}

see also: Decay.Q

Decay.Q — Method

Decay.Q(t, l2, l1; Q0=0, P0=1)

Calculate the the abundance of a radioactive daughter isotope Q in a decay-series after some time t, given its decay constant l2 and that of its parent l1. Optionally provide the initial abundances of the isotope Q0 (default = 0) and its parent P0 (default = 1).

After the derivation of Bateman, H. 1910. Solution of a system of differential equations occurring in the theory of radioactive transformations. Proceedings of the Cambridge Philosophical Society, vol. 15.

Q = \frac{\lambda_1}{\lambda_2 - \lambda_1} P_0 e^{-\lambda_1 t} + \left( \frac{\lambda_1}{\lambda_1 - \lambda_2} + Q_0 \right) e^{-\lambda_2 t}

see also: Decay.N2

Decay.R — Method

Decay.R(t, l3, l2, l1; R0=0; Q0=1, P0=1)

Calculate the the abundance of a radioactive daughter isotope R in a decay-series after some time t, given its decay constant l3, the decay constant of l2 of its parent Q, and the decay constant l1 of its parent's parent P. Optionally provide the initial abundances of the isotopes R0 (default = 0), Q0 (default = 1), and P0 (default = 1).

After the derivation of Bateman, H. 1910. Solution of a system of differential equations occurring in the theory of radioactive transformations. Proceedings of the Cambridge Philosophical Society, vol. 15.

Decay.act2conc — Method

act2conc(A, lambda, atomic_mass)

Convert a per gram activity to a concentration of radioisotope of interest, given its activity (as a [`RadUnit`](@ref)), its decay constant (a⁻¹, see [`Decay.λ`](@ref)), and its `atomic_mass` (in amu or g/mol).

Decay.conc2act — Method

conc2act(C, lambda, atomic_mass)

Convert a  concentration of a radioisotope to a Bq/g activity, given its concentration (g/g), its decay constant (a⁻¹, see [`Decay.λ`](@ref)), and its `atomic_mass` (in amu or g/mol).

Decay.dN2_dt — Method

Decay.dN2_dt(λ₂, λ₁; N2=0, N1=1)

Calculate the derivative of abundance of a radioactive daughter isotope in a decay-series given its decay constant λ₂ and that of its parent λ₁. Optionally give the current abundance of the isotope N2 (default = 0) and its parent N1 (default = 1).

Equation: dN/dt = λ₁ * N₁ - λ₂ * N₂

MathTex: \frac{dN_2}{dt} = \lambda_1 N_1 - \lambda_2 N_2

Decay.decayconstant — Method

decayconstant(halflife)

Calculates the decay constant (λ) from a half-life (t) with = log(2)/t

see also: halflife

Decay.halflife — Method

halflife(λ)

Calculates the half-life from a decay constant (λ) with log(2)/λ

see also: decayconstant