Decay

Documentation for Decay.

Decay.λ

Decay.D

Decay.D_

Decay.N

Decay.halflife

Decay.lambda

Decay.λ — Constant

Decay.λ

A NamedTuple library containing the decay constants and uncertainties of a selection of useful radiochronometric isotopes.

Decay constant central tendencies (typically means) are returned by providing the element (e.g. U) and isotope (e.g. 238), as in λ.U238. To call the 1σ uncertainty (standard error of the mean), give the same value with a σ prefix, as in λ.σU238.

References to be formatted:

238,235U Jaffey et al. 1971, Phys. Rev. C, doi: 10.1103/PhysRevC.4.1889 232Th ~~ Holden (1990) which is very near to the oft-used value of Le Roux & Glendenin 1963 (1.40 ± 0.007 Gy). Audi et al. 1997 (Nuc. Phys. A, doi: 10.1016/S0375-9474(97)00482-X) reports 14.05 ± 0.04 Gy and references Bonetti+ 1995, Phys. Rev. C, doi: https://doi.org/10.1103/PhysRevC.51.2530, which does not appear to report a λ_α. ²³⁴U, ²³⁰Th –- Cheng et al. 2013, EPSL, doi: 10.1016/j.epsl.2013.04.006 ²²⁶Ra –- Holden, 1990 (Pure and Applied Chem., doi: 10.1351/pac199062050941)

Decay.D — Method

D(t, λ; N=1, Do=0)

The classic age equation. Calculates the abundance of stable radiogenic daughter isotope D after time t, given the MODERN abundance of parent isotope N (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + N ( exp(-λ t) - 1 )

see also: D_

Decay.D_ — Method

D_(t, λ; No=1, Do=0)

Forward model a single decay system.

Calculates the abundance of radiogenic daughter isotope D after time t, given the INITIAL abundance of parent isotope No (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + Nₒ ( 1 - exp(-λ t) )

see also: D

Decay.N — Method

N(t, λ; No=1)

The classic decay equation. Calculate the remaining abundance of parent radioisotope N with decay constant λ after time t. Optionally provide an initial abundance No (default = 1).

Equation: N = Nₒ exp(-λ t)

Decay.halflife — Method

halflife(λ)

Calculates the decay constant (λ) from the halflife with log(2)/λ

see also: lambda

Decay.lambda — Method

lambda(halflife)

Calculates the decay constant (λ) from the half-life (t) with = log(2)/t

see also: halflife