Decay

Documentation for Decay.

Decay.λ

Decay.D

Decay.D_

Decay.N

Decay.N2

Decay.dN2_dt

Decay.halflife

Decay.lambda

Decay.λ — Constant

Decay.λ

A NamedTuple library containing the decay constants and uncertainties of a selection of useful radiochronometric isotopes.

Decay constant central tendencies (typically means) are returned by providing the element (e.g. U) and isotope (e.g. 238), as in λ.U238. To call the 1σ uncertainty (standard error of the mean), give the same value with a σ prefix, as in λ.σU238.

In most cases, I try to cite the original publications containing accepted measurements. In a few cases, I use compilations from reviews.

isotope(s)	reference	doi	note

²³⁸U, ²³⁵U	Jaffey+ 1991	10.1103/PhysRevC.4.1889

²³²Th	Holden 1990	10.1351/pac199062050941	[1]

²³⁴U, ²³⁰Th	Cheng+ 2013	10.1016/j.epsl.2013.04.006

²²⁶Ra	Holden 1990	10.1351/pac199062050941

⁴⁰K	Renne+ 2011	10.1016/j.gca.2011.06.021	[2]

[1] This value is similar to the often-used value of Le Roux & Glendenin 1963 (1.40 ± 0.007 Gy). Audi+ 1997 ( doi: 10.1016/S0375-9474(97)00482-X) reports 14.05 ± 0.04 Gy and references Bonetti+ 1995 (doi: [https://doi.org/10.1103/PhysRevC.51.2530, which does not appear to report a λ for α-decay.

[2] K40 and σK40 are NamedTuples containing fields β and e for the branched decay. The total decay constant is obtained with sum(λ.K40). I also include the now deprecated decay constants of Steiger & Jager 1977 (doi: 10.1016/0012-821X(77)90060-7). See also Renne+ 2010 (doi: 10.1016/j.gca.2010.06.017).

Decay.D — Method

D(t, λ; N=1, Do=0)

The classic age equation.

Calculates the abundance of stable radiogenic daughter isotope D after time t, given the MODERN abundance of parent isotope N (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + N ( exp(-λ t) - 1 )

see also: D_

Decay.D_ — Method

D_(t, λ; No=1, Do=0)

Forward model a single decay system.

Calculates the abundance of radiogenic daughter isotope D after time t, given the INITIAL abundance of parent isotope No (default = 1) and its decay constant λ. Optionally takes an initial daughter abundance Dₒ (default = 0).

Equation: D = Dₒ + Nₒ ( 1 - exp(-λ t) )

see also: D

Decay.N — Method

N(t, λ; No=1)

The classic decay equation. Calculate the remaining abundance of parent radioisotope N with decay constant λ after time t. Optionally provide an initial abundance No (default = 1).

Equation: N = Nₒ exp(-λ t)

Decay.N2 — Method

N2(t, λ₂, λ₁; N1o=1, N2o=0)

Calculate the the abundance of a radioactive daughter isotope in a decay-series after some time t, given its decay constant λ₂ and that of its parent λ₁. Optionally provide the initial abundances of the isotope N2o (default = 0) and its parent N1o (default = 1).

Equation: N₂ = (λ₁ / (λ₂ - λ₁)) * N₁ᵒ * (exp(-λ₁ * t) - exp(-λ₂ * t)) + N₂ᵒ * exp(-λ₂ * t)

MathTex: N_2 = \frac{\lambda_1}{\lambda_2 - \lambda_1} N^o_1 \left( e^{-\lambda_1 t} - e^{-\lambda_2 t} \right) + N_2^o e^{-\lambda_2 t}

Decay.dN2_dt — Method

dN2_dt(λ₂, λ₁; N2=0, N1=1)

Calculate the derivative of abundance of a radioactive daughter isotope in a decay-series given its decay constant λ₂ and that of its parent λ₁. Optionally give the current abundance of the isotope N2 (default = 0) and its parent N1 (default = 1).

Equation: dN/dt = λ₁ * N₁ - λ₂ * N₂

MathTex: \frac{dN_2}{dt} = \lambda_1 N_1 - \lambda_2 N_2

Decay.halflife — Method

halflife(λ)

Calculates the half-life from a decay constant (λ) with log(2)/λ

see also: lambda

Decay.lambda — Method

lambda(halflife)

Calculates the decay constant (λ) from a half-life (t) with = log(2)/t

see also: halflife